feat: Open All Bookmarks
As a user, I want to be able to open
all the bookmarks in a group with a
single click

This PR introduces a new button in the
group with which you can open all the
bookmarks in the group.

This is very handy, for users like me
who have to open set of dashboards,
apps at the beginning of a day.